Using explicit Balance type instead of int64

+++ code/go/0chain.net/smartcontract/storagesc/writepool.go
@@ -344,7 +344,7 @@ func (ssc *StorageSmartContract) createWritePool(t *transaction.Transaction,
 	}
 
 	var mld = alloc.restMinLockDemand()
-	if t.Value < int64(mld) {
+	if t.Value < mld {
 		return fmt.Errorf("not enough tokens to honor the min lock demand"+
 			" (%d < %d)", t.Value, mld)
 	}
